use uint32 for fee limit (e.g atoms/kB & sats/kB)

+++ client/core/wallet.go
@@ -28,7 +28,7 @@ type xcWallet struct {
 	hookedUp     bool
 	synced       bool
 	syncProgress float32
-	feeRateLimit float32
+	feeRateLimit uint32
 }
 
 // Unlock unlocks the wallet backend and caches the decrypted wallet password so
