+++ dex/networks/btc/config.go
@@ -45,7 +45,7 @@ type Config struct {
 
 	UseSplitTx       bool    `ini:"txsplit"`
 	FallbackFeeRate  float64 `ini:"fallbackfee"`
-	FeeRateLimit     float64 `ini:"feelimit"`
+	FeeRateLimit     uint64  `ini:"feelimit"`
 	RedeemConfTarget uint64  `ini:"redeemconftarget"`
 }
